Refactored frame into class
