+++ db/data.csv
@@ -3,3 +3,5 @@ id,username,password,balance
 86030215,hi,hi,0
 83564268,billy,password,0
 93578560,admin,admin,0
+99085440,new,password,0
+28258694,apple,apple,0
