+++ program.py
@@ -1,8 +1,7 @@
 import tkinter as tk
-import tkinter.messagebox
 import csv
 from random import randint
-from bElements import *
+from bFrames import *
 
 userDb = None
 
@@ -51,44 +50,7 @@ csvIO = dbIO('db/data.csv')
 root = tk.Tk()
 root.geometry("600x400")
 
-loginFrame = tk.Frame(root)
-
-userLabel = tk.Label(loginFrame, text="Username")
-userField = bEntry(loginFrame)
-
-pwLabel = tk.Label(loginFrame, text="Password")
-pwField = bEntry(loginFrame, show="*")
-
-def handleLogin():
-    result = csvIO.checkLogin(userField.get(), pwField.get())
-
-    if not result:
-        tkinter.messagebox.showerror(title="", message="Incorrect username or password!\nIf you have not yet created an account, press create one below.")
-    else:
-        userDb = result
-        tkinter.messagebox.showinfo(title="", message="Logged in successfully!")
-    
-    userField.delete(0, 'end')
-    pwField.delete(0, 'end')
-
-def handleCreate():
-    result = csvIO.createLogin(userField.get(), pwField.get())
-    if not result:
-        tkinter.messagebox.showerror(title="", message="That username is already taken. Please choose another.")
-    else:
-        tkinter.messagebox.showinfo(title="", message="Account created successfully!\nPlease press login.")
-
-loginButton = bButton(loginFrame, text="Login", command=handleLogin)
-createButton = bButton(loginFrame, text="Create an account", command=handleCreate)
-
-userLabel.grid(row=0, column=0, pady=2)
-userField.grid(row=0, column=1, pady=2)
-
-pwLabel.grid(row=1, column=0, pady=2)
-pwField.grid(row=1, column=1, pady=2)
-
-loginButton.grid(row=3, column=0, columnspan=2, pady=2, sticky="ew")
-createButton.grid(row=4, column=0, columnspan=2, pady=2, sticky="ew")
-
+loginFrame = LoginFrame(root, csvIO)
 loginFrame.place(relx=0.5, rely=0.5, anchor="c")
+
 root.mainloop()
